+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Version History
 
+## Unreleased
+
+* [THORN-533] Implemented a pasdBus health model. This exposes attributes e.g. smartboxFailedPollCount to keep track of the number of failed polls for each device. The behaviour is configurable using the following new device properties:
+  - FailedPollWindow: sliding-window length (seconds)
+  - FailedPollPruneInterval: how often (seconds) to prune expired entries from the failed-poll window
+
 ## 6.2.0
 
 * [SKB-1204] Only set DevState to ALARM when the device is ON.